:construction: chore(timermanager): updated the calling and fix building error

+++ server/nyxcore/core/nyx_timer_manager.h
@@ -61,7 +61,7 @@ class timer_manager : private boost::noncopyable {
   __timer_set timers_set_;
   mutable std::mutex timer_mutex_;
   py::object func_;
-  static std::int64_t tick_interval_;
+  static std::size_t tick_interval_;
 public:
   timer_manager(void)
     : func_(py::detail::borrowed_reference(Py_None)) {
@@ -70,7 +70,7 @@ public:
   ~timer_manager(void) {
   }
 
-  static void set_tick_interval(std::int64_t interval) {
+  static void set_tick_interval(std::size_t interval) {
     tick_interval_ = interval;
   }
